radicle-cli: Use `rad remote` to list rad remotes
Help users by providing `rad remote` as a utility to list and operate on
radicle remotes.

Repurpose an unused method `radicle_cli::git::remotes()` to do what its
comment suggests, list a project's radicle remotes.  Use it to list the
remotes to the user.

An example file is provided to demonstrate the commands use.  It is
intended to be temperary until the command is completed and the
documentation better incorporated into the project.