cli: Store refs when interacting with rad node is needed
During interactions with the radicle cli, certain operations
require communication with the Radicle node.

To accommodate scenarios where the node is down or the user
prefers working offline, this commit stores relevant information
about the refs to be announced.

This way, at restart time, we can reannounce the stored refs.

Additionally, this commit introduces logic to sync with the remotes
after a patch. Previously, this step was not automatically
performed, necessitating the manual use of `rad sync`.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>