node: Reannouncing the refs stored inside the database
This commit adds the possibility to check the database at startup
for not annunced refs and reannounce them to the respected peer.

The only problem with this is that we do not have a way
to connect with previous peers (except if the radicle-node is not
configured to do so). As a result, this logic may also be need
when we connect with a node, or we need to store the node where we
are connected and reconnect before reannouncing.

I would prefer the first option because it seems more scalable.

Link: https://app.radicle.xyz/seeds/seed.radicle.xyz/rad:z3gqcJUoA1n9HaHKufZs5FCSGazv5/issues/278ed198530fd701ff4c977091e7d829c7d98cc8
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>